refactor: 포토톡 갤러리 모달 경로 수정
