+++ src/components/common/PhotoTalk/Gallery/PhotoTalkGalleryPreview.tsx
@@ -1,5 +1,5 @@
 import PhotoTalkGalleryGrid from '@/components/common/PhotoTalk/Gallery/PhotoTalkGalleryGrid';
-import PhotoTalkGalleryModal from '@/components/common/PhotoTalk/Modal/GalleryModal/PhotoTalkGalleryModal';
+import PhotoTalkGalleryModal from '@/components/common/PhotoTalk/Modal/PhotoTalkGalleryModal';
 import { phototalkData } from '@/constants/phototalkData';
 import { UserMode } from '@/types/users';
 import { useState } from 'react';
