+++ src/components/common/PhotoTalk/Modal/PhotoTalkGalleryModalFooter.tsx
